Hopefully solve bug #549, "The resulting page after writing a comment is not intuitive" by providing a link to the article commented.
Invalidation of the textstat cache is done to ensure that the comment is shown
in the list of comments - if not, the user will probably don't think the
comment was submitted. However, the user might see his/her own comment if
using "read next comment". Perhaps that function should be "read next unread
comment" instead?